[map] increase dynamic range of map colors (#1096)
domain and range must match -- we were setting domain only with extent, but range with 3 values, so the third (darkest color) was ignored. increased the range, making it darker at the end, using the max color for the median instead.

also fixed the legends to always include the min and max values.