[Disease Browser] Add disease chart descriptions, common name and integrate interfaces in one file  (#1727)
In this PR, I've added the description for each chart and table. In
addition, I've linked it to the graph browser view and added the disease
common name as the first header (instead of the disease ontology ID,
with the former being more intuitive and easy to understand). I've also
added a types file with all disease browser interfaces added to it, for
the sake of better organization of the code.
<img width="1051" alt="Screen Shot 2022-08-22 at 5 05 51 PM"
src="https://user-images.githubusercontent.com/57412795/186033657-ef74b7ee-ba73-43c5-8f0a-b8a3b06be0cd.png">

Co-authored-by: Suhana Bedi <suhanabedi@google.com>