[map refactor] Add data fetcher modules and a chart store (#1775)
This is a no-op change to existing code and prepares for future updates
of chart_loader.

The goal is to remove the dependency of data fetch and processing.
Ideally, the tool only needs to fetch the needed data and will render
the map when all necessary data is ready. The other data can be loaded
on the background or on-demand, which can be controlled independently
without touching the component logic.

Some design thoughts:

- Each data load should be in a separate module, which fetches data and
sends the raw data to the desired component. These modules are put in
/fetcher folder. New data fetcher should be created in the same pattern.

- All the raw data are used in various ways to construct the chart data,
the construction depends on context (user input) and is dependent on
data availability. Here I created a ChartStore to hold all the raw data
and also the context when they are created. This has the following
benefits:
1. Allow each fetcher to independently "dispatch" the data. The data is
merged using "useReducer" hook in the UI component.
2. The UI component can check if the data is ready by comparing the
context in ChartStore vs the current context.
3. In future change, the async data processing dependency can be
resolved by "useEffect()". By checking the readiness of the data, each
useEffect() can exit early if needed.