[NL Interface] Correlation Detection - First Version (#1971)
This PR adds:

1. Basic clustering-based correlation detection. It mostly only works
when the query matches two very distinct set of SVs, e.g. "median income
and auto theft California".
2. The Detection object has all the attributes set.
3. Increased the number of SV matches we get from the embeddings matcher
(from 10 to 20). The reason is that we now have a growing number of
sentence alternatives for each SV and as a result the matching produces
many duplicates.
4. For the debugInfo rendered on the website, I have concatenated
everything in a string. It is not the best rendering but for now, actual
detection is perhaps more important than the formatting of the
debugInfo. See attached image for now the debugInfo looks like:

<img width="1033" alt="Screenshot 2023-01-08 at 5 58 19 PM"
src="https://user-images.githubusercontent.com/1021616/211231670-9652a7e7-5519-47a7-bb3e-f761b4a20461.png">