Add a script to manually verify the flow of NLP features (#2007)
Since the flow is fixed and we have fast merging PRs, it would be nice
to have a way to verify the flow end to end with a script. This PR only
verifies Palo Alto is correctly detected. The rest of the flow will be
followed.

The goal is to test: query -> nl/data API response, with contexts in
between.

Co-authored-by: Alex Chen <alexyfchen@google.com>