[NL Interface] Replacing SV search and NER place search with the new server API endpoints (#2122)
This PR replaces the previous SV search (on embeddings) and NER place
search with the new NL models server API calls.

Note: for now, the dependence on the embeddings models and builds is not
completely removed. This will need one more API end point first
(returning the embeddings vector for individual queries). This will be
done in the next PR.