Fork ranking into by place and by var. Also improve fallback (#2157)
Bo had the right intuition earlier today to break up ranking into by
places and vars. With the addition of "time delta" for ranking across
vars, this is better.

Also, often times when the ranking, correlation and contained-in fails
on a SV a better choice is to show the place rather than attempt some
poor fallback of population in county, which invariably also fails.