[NL Interface] Removing all "ML" classifiers and training etc (#2248)
This is being done as part of the NL Tasks. Here are some reasons:

1. Several of the ML classifiers are not being used and use up time to
run through training at startup. No need for that.
2. The only ML classifier that was actively being used was the
"contained in" classifier but even in that case the specific type of
classification was being narrowed down using some string matching
heuristics. Therefore, the classifier wasn't adding much value and it
was thought to simplify things (discussed with @pradh ).
3. Just to simplify things--if we need to bring back ML
classifiers/attribute labelers later, we can do so easily without the
existing burden of the ML classifiers.

Tested locally, all unit tests pass and the demo script works fine. If
testing this locally, please remember to delete your local cache first.