[NL Interface] Stopwords removal improvement (#2255)
This PR makes a couple of (necessary) enhancements. The list of stop
words now includes sentences and regex expressions (this has happened
over time because various query attributions/classification look for
words and sentences). This PR does the following:

1. Instead of splitting every stop word entry (which could be more than
one word), look for the entire entry before finding a match and removing
those words. This has the advantage of not removing unnecessarily. For
example, for the query "How big are the public elementary schools in
Sunnyvale", current the word "public" is also removed but that should
not happen. Only "elementary schools" should be removed (along with
"how" "are" "the" "in").

2. updating to regex based removals allows several regex expressions to
match (which was not happening today).

Added some new unit tests + tested this on the current and past demo
scripts.