Track SV strings in github and add index building code (#2341)
As part of updating the index, devs show now also send CSV diffs so we
can know what exactly is changing.

The `build_embeddings.py` is a straightforward port from the [notebook
here](https://colab.sandbox.google.com/drive/1s2qYkbjZ6ZXoDfYTWWkk5SMgEVQFBAtK#scrollTo=lzmyZYF9C5wv).

Note that the canonical version is still the spreadsheet. I originally
wanted the checked-in file to be that, but the current sheet has some
funky string concat across the descriptions that should rationalize
first.

We might also need more stuff here as we scale to 10K SVs, so this is
just an incremental improvement.