Store mixer-api-key in the website GCP project instead of in the mixer project (#2422)
This aligns with the regular practice that api key sits in a separate
(client) project. With this, custom DC users can store the API key in
the custom project.

Also make the change that local custom DC dev talks to a local mixer
instead of GCP mixer. GCP mixer per custom instance is not possible to
maintain and usually is out of date. This is especially true for a 3p
customer.