[NL Interface] Renaming model.yaml to embeddings.yaml and associated changes (#2875)
Fixing a historical artifact where `model` got confused with
`embeddings`. Now that we have fintuned models, we will need to keep the
distinction clear. This PR is the first step along the way where
`model.yaml` is renamed to `embeddings.yaml` and the `tuned_model` is
moved out from this file to a separate `tunedmode.yaml` file.

There are other downstream changes which were impacted by this update.
This PR should have all those changes.

In the next PR, the fine tuned model will be integrated in the server
and nl_server (wherever the base model was being loaded).