[NL Interface] Updates to Finetuning, Embeddings, SV Differ (#2896)
This PR does the following:

1. Introduces the ability to finetune a final model without using an
intermediate step (BASE -> FINAL). This is found to currently be more
useful than the intermediate step (upon which the investigation can
continue).
2. Support building embeddings based off a local model path (this is
useful when experimenting with a new finetuned local model which isn't
yet checked in to PROD).
3. Enabling the SV Differ to take a different model (finetuned model)
for the test embeddings.

Note: this PR does not update the embeddings or model yaml files. A
sample differ report for the finetuned model from (1) is
[here](https://drive.google.com/file/d/1mgsIIUZseGQx2mClZkps_dGqj6B2SVS8/view?usp=drive_link).
Note that this finetuning is BASE+FINAL (and does not have the
intermediate tuning step). The differ report is produced vs the existing
PROD (small) embeddings (which used the base model). It will remain a
work in progress with some iterations.