LLM: Improve LLM fallback (#2895)
Generally, if a query can likely be handled using heuristic, we do that.
Only rarely do we fallback to LLM, and when we do, we are explicit about
what from LLM-based detection we actually use (only place, only
variable, or both).

Concretely, based on what the heuristic approach first detects, here are
the rules for LLM fallback:

1) If no SV, SIZE and EVENT classification was detected in this query,
and no SV in the context too, fallback to LLM for SV.

2) If no place and COUNTRY type was detected in this query, and no place
in the context too, fallback to LLM for place.

3) For a multi-SV query (that we're sufficiently confident is multi-SV),
fallback when we have >2 SVs, or if there's no correlation/comparison
(because [poverty vs. asthma] queries we can handle better with
heuristics), fallback to LLM for SV.

Also, refactor the detection routing logic to a separate file instead of
in `api.py`.