Pure refactor of classifiers and place detection (#2916)
This drops the whole `Model` object in services and uses direct calls to
classifiers and place detection. Also, factor out some place detection
logic that is currently in `llm_detector` and `heuristic_detector`.

This is to make the follow up Place Detection via DC APIs simpler.

NOTE: The test change is because of World Bank refresh I think.