NL: misc fulfillment fixes (#2923)
#### 1 - Handle mixed fallbacks better

If we retrieve a place or variable from context and also do place
fallback, provide a unified message.

Before this PR we would not report place fallback when we fallback from
a place in context. [asthma in california] -> [what about blood
pressure?]

Now we'll say, `Tried looking up relevant statistics for "California"
based on your prior queries, but found no results. Here are results for
"United States" instead.`

#### 2 - Avoid duplicate SVs in comparison bar charts

[How long do workers in Maryland live] from earlier today

#### 3 - Avoid bar charts that have no SV in title


![image](https://github.com/datacommonsorg/website/assets/4375037/243a5c3a-2777-48e5-962d-2f829b0a2607)

These come about because we were expanding a "main SV" even when it
doesn't pass existing check, so fix that.

#### 4 - Add fallback info debug page


![image](https://github.com/datacommonsorg/website/assets/4375037/aa8e48ff-49f3-4582-9c75-68b3acb2412c)

#### 5 - Exclude more per-capita SVs

This is a temporary thing until we have [chart controls for
per-capita](https://github.com/datacommonsorg/website/issues/2917).