[NL Interface] Fixing the local cache bug (#2931)
We are now exceeding the local cache default size (1Gb) which meant that
the local development caching was not working. Fixing this issue by
increasing the cache size.

This PR also ensures that if an Embeddings object was not found in the
local cache (used in the server/embeddings_test.py) then the finetuned
model is downloaded/loaded and associated with the embeddings index in a
new Embeddings object.