[NL Interface] Updating medium_ft model + medium_ft based embeddings (#2955)
In this PR, we have the following:

1. Lots of updates to the base SV names and descriptions
2. Updates to the PaLM alternatives
3. Few updates to the basic fine-tuning pairs (based on some bugs/bad
behavior seen recently)
4. Rebuilding of the finetuned model (medium_ft)
5. Building a new medium_ft embeddings index using the model in (4).

The SV Differ report (comparing the new medium_ft index vs the existing
medium_ft index) is here:
https://drive.google.com/file/d/1yUnl1a4DYf_C96Y7Y5fpnZJTBXK4rnp0/view?usp=drive_link

It mostly looks like a strict improvement. Most of the top SVs found are
the same. But some clear improvements based on fine tuning and SV
description edits/improvements. The most noticeable differences are in
the various emissions SVs which seem to have better behavior now instead
of always matching to the Sources topic.