Update Equity and Sustainability explore examples (#3106)
Updates the explore pages for equity and sustainability with @rvguha's
brainstormed list, with newly curated topic DCIDs in the URLs. A few
queries/places were tweaked to match the data and place types we have
available.

Also copied sources and variable count from demographics to the equity
section because there's a lot of overlap in SVs. We can curate sources
and update variable count in a followup PR.