Add Explore Chart Layout, based on a flag (#3259)
This change adds a new fulfillment handler called "basic" which
encompasses:
* simple
* containedin
* ranking_across_places
* ranking_across_vars
(the names need some tweaking, like what's basic vs. simple, but that's
in followup)

The "basic" handler, depending on whether it is explore vs. chat, will
route to one or more of these components appropriately.

Also, for every single-line timeline (in both modes), add a highlight,
and demote a single-bar barchart to highlight.

IMPORTANT NOTE: This change makes all of the existence checks to require
"is it single point", so that makes things a LOT more expensive, esp for
[What are the projected temperature extremes across california].

REGRESSION NOTES:

1. After this change, chat will show a bunch more charts. As an example,
for query [asthma in california counties], suppose it returns asthma and
childhood asthma, and suppose asthma doesn't exist at county-level (it
does exist, but just suppose), then today we'll not show asthma at all.
After this change we'll show asthma line chart and childhood asthma map.

2. There are cases where the "extension SVs" maybe removed (in diffs),
that will be fixed in upcoming PRs.