[NL Interface] Updating the SDG Index (#3337)
Adding "sdg" and "sustainable development goal" to the SDG goals and
indicator names as alternatives in the SDG index. This helps match to
the right SSDG goal/indicator when asking for things like:

"Compare SDG goals on climate change between Brazil, China and India".

Previously, we saw this:
https://screenshot.googleplex.com/857NpzMbggWKU68 (the word "SDG" was
matching to other texts in the index but since all indicators have "sdg"
in their alternatives, that could match to more random SDGs). Now, with
every name also having an "sdg" in the name (as another alternative), we
normalize the effect of "sdg" in the query. The updated matching is now
like this: https://screenshot.googleplex.com/68xeyoZhUhugU6M and the
matching index matching to SDG_13 at the top which is the right thing.

Another example: "progress on SDG on poverty in India and Zimbabwe" also
matches to the poverty SDG goal at the top:
https://screenshot.googleplex.com/9cwSEiRT8y4KheK