add sdg chart titles (#3745)
- add sdg chart titles to the cached chart titles
- updates chart titles to include a list of blocklisted dcs
- this is needed because we want curated titles for sdg svs on the main
dc because otherwise titles on the explore page will have things like
'2. No Hunger' which is confusing when there's no context into what the
number is for. On the other hand, for the sdg static site, the UN wants
all the sdg goal/target/indicator numbers and the exact original names.