Update secret and api key handling for local | custom development (#3754)
Currently local.Config should be for DC internal usage but it is
inherited by custom DC as well. This makes it a general local config and
use local.DCConfig for base DC development.

Also requires DC_API_KEY to be set for local development and throw an
error to message that.

Rename all "MIXER_API_KEY" to "DC_API_KEY" for consistency.