NL: Fix 3 fulfillment bugs (#3753)
This PR fixes the following bugs.

**1. b/303103114: Presence of a superlative keyword can break other
query-shapes**

* Ensure that superlative is ranked below time-delta, etc. So if a query
has time-delta and superlative, we prefer the former.
* Also, in `fulfill()` handle failure return from `set_overrides` by
looping over the query-types

Before:

![image](https://github.com/datacommonsorg/website/assets/4375037/4fd823d9-e35f-4fb3-a7e6-db369bb96c65)

After:

![image](https://github.com/datacommonsorg/website/assets/4375037/545683c9-f291-414d-8110-ad3a8be123f0)

**2. b/307806796: Avoid incorrect X vs. Y titles when correlation
queries have no data**

* To infer the page topic title, we use the `orig_svs` in the ChartVars
(which can be a topic and differ from the plotted SVs). For correlation
queries there are 2 `orig_svs` entries. When there is a no-data error,
we would end up showing only 1 SV, and in that case, we need to be able
to tell which among the 2 `orig_svs` we should include for topic
display.
* Change `orig_svs` to be a map, with the plotted SVs as value.

Before: 

![image](https://github.com/datacommonsorg/website/assets/4375037/c2ef9d96-aa50-4fe9-82b1-ef43c337e5df)
After:

![image](https://github.com/datacommonsorg/website/assets/4375037/cdbe39b3-cce3-49f7-b33d-9e0bb9ca7c50)

**3. b/308208124:  SVs in topics do not show up for a few query-types**

* We fail if the ChartVars has more than 1 SV. Instead also check that
it is part of peer group. So a useful SV (like Median Income) part of a
topic will show up.
* This affects 3 query types: TIME_DELTA_ACROS_PLACES,
FILTER_WITH_SINGLE_VAR and FILTER_WITH_DUAL_VAR
* For these, we only show one SV, and that stays the same.
* In these handlers, it is also important to reset the ChartVars SVs to
the first SV plotted.

For income, we show a lower-ranked standalone SV not part of a topic
now, but with the change we show median household income that's part of
a topic (with other SVs).

Before:

![image](https://github.com/datacommonsorg/website/assets/4375037/d27f6ccb-b3a4-4fa9-b4ff-2ce99f93cad1)

After:

![image](https://github.com/datacommonsorg/website/assets/4375037/c335acaf-ca7d-4588-aece-b5887c78f164)