Suppress low-confidence SVs in `strict` mode (#3768)
Currently, we have a low-confidence threshold (below which we print a
message) of `0.7`, and a drop threshold (below which SVs are dropped) of
`0.5` for SVs.

The main change in this PR is, for`strict` mode, we increase the drop
threshold to `0.7`. This is done by plumbing in the "threshold" value to
the NL server, depending on the setting of `mode=strict`.

Importantly, the `strict` mode is:
* supported only for `api/data` API that nodejs uses
* incompatible with LLM.

Concretely, setting of `mode=strict` implies: `sv_threshold=0.7`,
`detector=heuristic`, `use_default_place=False`

Additionally,
* Deprecate `udp=` param
* Add support for `mode=strict` to NL app, since that exercises the
`api/data` flow
* Move tests from `explore_test.py` to `nl_test.py`, also for the same
reason.