[vis tool] fix source selector on timeline (#3889)
This fixes the source selector for the timelines on vis tool
(https://screenshot.googleplex.com/kX3735zMTDF7uFR). The problem was
that the endpoint for fetching the facets enforced strict slashes, but
the client was not passing in a slash and so the client was being
redirected to a location it could not access.