fix autopush deployment (#4020)
Update autopush website-app to deploy in default-pool. This fixed
deploying autopush when trying manually

Background:
- autopush deployments have been failing since website cron tests have
been added
- there are 2 different node pools used in the autopush instance, one
with 4 CPU and one with 8 CPU
- I noticed that a lot of website-app pods were deployed in the 8 CPU
node pool and they seemed to have a lot of restarts. In a past
experiment, we've learned that website-app pods have trouble starting in
node pools that have more CPU than 4.
- So I tested moving all the cron jobs to the 8 CPU node pool and all
the website-app pods to the 4 CPU node pool and it seemed to work. (Will
see how well it does once this PR is in though)