[NL] Update timeline date range to work with FE change (#4111)
There was a FE change for how timeline tiles handle date ranges.
Timelines used to cut dates to the lowest granularity between specified
date range and observation date before comparing them, but now timeline
will only cut observation dates to the granularity of the specified date
range before comparing dates. On the NL side, we want the original
behavior so that if a start date like 2020-04 is specified, observations
from 2020 still show. This is a change to the date range we pass in the
tile config from the NL so that we get the original behavior.

Additional change:
- update the regex for date recognition because query will actually be
in lower case by the time we get to the date parsing code
- make sv_place_facet and sv_place_facets consistent to all be singular