Revert "Upgrade d3 to v7.4" (#4196)
Reverts datacommonsorg/website#4182

The node js can not compile due to this change due to d3 updated to es
module (instead of being common js module). The typescript compile and
import becomes very tricky.

Should come back to this once we have a better understanding of the
module and imports.