Major Index Update: change default index to base_uae_mem (#4241)
This is a major release of the NL feature to replace the all-mini-lm-l6
model with uae-large-v1 that is hosted on vertex AI.

Also made fundamental improvements to stat var descriptions, and only
use one accurate description per stat var. This get rid of the need to
use alternatives.

This adds some small debug info UI improvement.