Miscellaneous updates and improvements (#4673)
## Description

This PR contains miscellaneous updates and improvements triggered from
comments in an earlier PR:
[4599](https://github.com/datacommonsorg/website/pull/4599)

It contains the following updates:
* The TypeScript breakpoint object is standardized to defaults, and
comments now reference the TypeScript and SCSS version so that they can
be kept in sync with further changes.
* The `Topics` component has been generalized into a `LinkChips`
component, making it a more reusable content component. This update was
triggered by the request to change the CSS class from "topics" (which
was also done along with this larger change). We can generalize other
content components in later PRs.
* The directory for the SCSS files relating to content components is now
renamed to make that structure more clear.
* Updates to comments, copyright notices and the variable import file
name.