Adds Continents to AutoComplete (#4693)
This PR implements a hack to add autocomplete for continents. We always
fake that Continents came back in the response from the Google Maps
Predictions API response, and hardcode the Place ID to DCID.

The Place IDs were fetched from here:
https://developers.google.com/maps/documentation/places/web-service/place-id

For one word continent, we just add a matching for the single word
query.
For two word continents (S|N America), we add matching for the single
word query, and two word query if applicable.

We then compute the match score, and remove all continents with positive
scores (negative score means a better match).

Note that we prepend the continents in order to give the continents an
edge over other places that have the same score.
I also added a boost to scoring if the query matches the response from
the start in order to give "America" an edge over "North America" when
typing "Ameri".