Adds GA even trigger for when autocomplete gets triggered. (#4694)
Trigger a Google analytics Event whenever we trigger auto suggestions.

This only gets triggered once per series of suggestions so as not to
falsify the data. The aim of this event is to be compared with the
number of times the autocomplete suggestions is accepted. That event
already gets triggered.