Global header search (#4740)
### Description

The purpose of this PR is to have the search bar in the header on all
pages.

Previously, pages by default did not have the search in the header, and
had to be flagged in order to provide the search bar. Now, this flag is
reversed and pages will contain the header search by default.

The search bar has been removed from the explore page, and the more
complicated search functionality provided in that page (such as the
ability to programmatically change the current query and provide a debug
modal) is now available to the universal search bar in the header.

Related layout updates have been made to ensure that the explore page
displays correctly with the internal search now gone, and the debug
"bug" button is now aligned nicely (with the global search at the top).
The debug modal has also been revamped.

### Notes

When the search bar was inline in the explore page, that page could
interact and share state with its inline search. This allowed the
explore page to provide and programmatically update the query string as
well as to provide debugging information, among other things.

We now have to communicate that information from the explore page all
the way to the header. Normally, we might consider options like React
context, or lifting of state.

However, in this case, the header and the explore page are in totally
separate React apps with separate mount points, and that limits our
ability to use more traditional methods to communicate this data.

The solution implemented was to create a global "Query Store", that can
be updated by a component in one app, and then have those changes
provided to a component in another app. The query store provides a hook
to facilitate this communication. This allows these otherwise completely
independent applications to communicate with each other in a reactive
way.

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>