[Place Page Revamp] Add logic for filtering charts (#4848)
Implements the fix for removing charts that have no data on either of
the cases according to their block scope.
So either for the current place, the child places, or the peer places
within the parent.

Modifies the call to relatedPlaces such that it comes BEFORE placeCharts
(as opposed to in parallel). We have to do this in order to pass in the
parent place dcid to the placeCharts method so that we can verify we
have data for peer places within parent.

With this change, all charts that have NO data are removed.