[Fixit Fridays] Make webdriver helpers wait till the element exists, and use those methods in tests (#4862)
Modified the base_utils find_elem and find_elems methods such that they
first have to wait for the element if it's not found. This allows us to
remove all that logic from within each test to check if the element
exists before trying to find it.

Also fixed how the XPath is sometimes used as the entire path multiple
times in the same test.

Moving forward, we should aim to always used find_elem, and find_elems
to retrieve elements for webdriver tests.
Will keep working towards moving over the old tests to using this
approach.

This also fixed the drought webdriver test which was previously skipped
for flakiness. I expect that this will allow us to unskip many other
tests as well.