Data Sources (#4864)
## Description

This PR implements the tabbed data sources pages.

To facilitate this, we have developed a generic tab component that is
used here and can be used elsewhere.
The tab component can optionally be “routed” or “standard”. We are using
routed tabs for the data sources.
With routed tabs, when you select a tab (such as “demographics”) then
“demographics” will appear in the url. For example, with the base page
“/data”, the demographics tab will be “/data/demographics”. If the user
arrives directly on the page “/data/demographics” (for example, from a
link or a bookmark), then the demographics tab will be selected by
default.

Standard tabs (available within the tab component, but not used here),
work very similarly. However, the URLs are not altered, and the tabs are
handled entirely by internal state. These could be useful if, for
example, you have multiple tab components on a page, where it may not
make sense to put the tab on the URL.

### Tab Design

The look and feel of the tab design was inspired by Material Design and
tab sets found on Google pages. Following this, we made the tabs
responsive by adding navigation arrows dynamically when the tab set
container can no longer contain all the tabs. Users can navigate the
tabs by either clicking those arrows, or on a mobile device by swiping.

### Notes

The data used from the data sources come from the following pages in the
docs:
https://docs.datacommons.org/datasets/Energy.html (where “Energy” can be
swapped out with “Health”, etc). In order to get this information in a
form that can be used in the frontend, we have provided a Python script
that parses the relevant doc pages and creates the json that we then
use.

This measure may be temporary. In the future, the same scripts that
generate the doc pages (located outside of this repository) could also
generate the json used by the data page implemented here. However,
pending the that implementation, the provided script here can update the
content of these data pages.

The script does not need to be run to test or merge this PR. It has
already been run and the output has been committed.

We only need to run it if we wish to update the data based on changes to
the docs. Instructions on running the script are provided in a README
file in the same directory (/tools/data_sources/README.md).

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>