Change sample questions (#4896)
Which country has the most old people? --> Doesn't show very good data
to highlight since we mainly only have US data, until the 5th chart
which is life expectancy and that sv is highlighted in one of thehealth
question.
I changed that question to "Tell me about age distribution in the USA."

Change GDP highlight from iran to new zealand.

I also replaced the language highlight from Tamil to Spanish since in
the US, the Tamil speaker count is so much lower that the data is a bit
less significant. Max state was at 0.3% of population compared to
multiple states with 20%+ Spanish speakers.
I couldnt get any data for Tamil speakers by indian state.

I also removed immunization in Pakistan since we don't have that data
and I couldnt find any other country for which we had immunization
rates. Perhaps that was from one of the datasets we had to remove for
licensing?