[Dynamic Placeholder] Improve Google Analytics for dynamic placeholders (#5045)
Make a new event for rendering the search bar with dynamic placeholders
since the text value matters on whether it's enabled.

This way we can actually compare the # of events for the search bar
being rendered , and the number of events being rendered with dynamic
placeholders. And we can compare that ratio to the number of
explore_search_q events.

This also propagates whether the dynamic placeholders are enabled all
the way to the onSearch trigger such that we don't miscount when the
placeholder is hidden due to the query.