Place Data Overview Page (#5094)
## Description

This PR implements an initial version of the place data overview page,
as referenced on approximately page 66 of the go/dc-rolling-ux
slideshow.

The primary purpose of this PR is to create and prepare the various
components that make up the page. However, the initial framework of the
page is also created.

## Notes

### Feature Flag

The availability of the routing for this page is determined by a new
feature flag: `place_data_overview`, set to true only in `local`.

### Routing

The base route is `/data/place/<place dcid>`. Canada can be accessed at
`/data/place/country/CAN`

### Data

Pending the availability of actual information, this page currently uses
mock data for Canada (as per the dc-rolling-ux example). Other places
will not be available (unless they are also added to the mock data).
Once the source of available information is determined, this can be
replaced with a fetch of the actual data for the page. Typing is
provided for this data (which may be altered to fit once the real data
is available).

### Components

Some of the basic components (link_chips, section) have been enhanced to
allow for the kind of customization required here. As part of this,
we've introduce `sx` props - inspired from Material UI (MUI) - that
allow consumers of these components to freely override the styles of
those components. We've also added a generic box component.

Then there are the specific components used on this page. These should
be able to be used with minimal alterations once the real data is
available.

### Stat Var Hierarchy

Because of the inclusion of the StatVarHierarchy component, we also need
to include SCSS associated with this page, as well as the Material
fonts. We could put the StatVarHierarchy on the list of components to be
converted to Emotion, which would allow it to be used here and in other
places without bringing in these other assets.

## Testing

This page can be viewed at: http://localhost:8080/data/place/country/CAN

## Screenshots


![data_place_country](https://github.com/user-attachments/assets/dde37266-432f-44a4-9214-aecc98431ee9)

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>